reorganize the incident information files
[#152464352]
