+++ app/javascript/selectors/screening/incidentInformationShowSelector.js
