+++ app/javascript/views/IncidentInformationShow.jsx
